Final code to pass off on the project
Updated comments and final code that will be passed off to the MSU bio inspired lab. Any change in code or failure to use code committed by me, Jaiden Westover @jwwwest, and the subsequent consequences of such actions is the fault of the user, the High speed actuator group will not be liable for such consequences.